fixed style things
