+++ src/HIVCohortFilter/HIVCohortFilter.css
@@ -212,12 +212,12 @@
 
 #ec-counts-overlay-1 {
   top: 372px;
-  right: 410px;
+  right: 387px;
 }
 
 #control-counts-overlay-2 {
   top: 372px;
-  right: 55px;
+  right: 32px;
 }
 
 #download-EC-cohort-overlay {
